requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.8
joblib==1.3.2
tensorflow==2.15.0
pillow==10.1.0
python-socketio==5.10.0
eventlet==0.33.3
